Refactor services and remove legacy logic
